+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Version 2.0.0
 * Use `geom_type` property instead of deprecated `type`
 * Add the possibility to give a coordinates transformer
 * Add a `geojson` option. See [#107](https://github.com/tilezen/mapbox-vector-tile/issues/107)
+* Refactor the options using the `per_layer_options` and `default_options` dictionaries.
+* Add the option `max_geometry_validate_tries`.
 
 
 Version 1.2.1